Rename 'image' to 'platform'
