+++ crates/notion-core/src/lib.rs
@@ -33,10 +33,10 @@ mod distro;
 pub mod env;
 mod event;
 pub(crate) mod fs;
-pub mod image;
 pub mod manifest;
 pub mod monitor;
 pub mod path;
+pub mod platform;
 mod plugin;
 pub mod project;
 pub mod session;
