+++ src/command/deactivate.rs
@@ -1,4 +1,4 @@
-use notion_core::image::System;
+use notion_core::platform::System;
 use notion_core::session::{ActivityKind, Session};
 use notion_core::shell::{CurrentShell, Postscript, Shell};
 use notion_fail::{ExitCode, Fallible};
